Addressing issue #71: more than 256 body literals

+++ include/vlog/ruleexecdetails.h
@@ -17,9 +17,9 @@ struct RuleExecutionDetails {
     std::vector<RuleExecutionPlan> orderExecutions;
 
     std::vector<uint8_t> posEDBVarsInHead;
-    std::vector<std::vector<std::pair<uint8_t, uint8_t>>> occEDBVarsInHead;
+    std::vector<std::vector<std::pair<int, uint8_t>>> occEDBVarsInHead;
     std::vector<std::pair<uint8_t,
-        std::vector<std::pair<uint8_t, uint8_t>>>> edbLiteralPerHeadVars;
+        std::vector<std::pair<int, uint8_t>>>> edbLiteralPerHeadVars;
 
 
     RuleExecutionDetails(Rule rule, size_t ruleid) : rule(rule), ruleid(ruleid) {
