+++ include/vlog/seminaiver.h
@@ -91,7 +91,7 @@ class SemiNaiver {
                 size_t limitView,
                 std::vector<size_t> &cards);
 
-        void processRuleFirstAtom(const uint8_t nBodyLiterals,
+        void processRuleFirstAtom(const int nBodyLiterals,
                 const Literal *bodyLiteral,
                 std::vector<Literal> &heads,
                 const size_t min,
@@ -100,7 +100,7 @@ class SemiNaiver {
                 const bool lastLiteral,
                 const size_t iteration,
                 const RuleExecutionDetails &ruleDetails,
-                const uint8_t orderExecution,
+                const int orderExecution,
                 std::vector<std::pair<uint8_t, uint8_t>> *filterValueVars,
                 ResultJoinProcessor *joinOutput);
 
@@ -204,7 +204,7 @@ class SemiNaiver {
             return chaseMgmt;
         }
 
-        virtual FCTable *getTable(const PredId_t pred, const uint8_t card);
+        virtual FCTable *getTable(const PredId_t pred, const int card);
 
         VLIBEXP void run(size_t lastIteration,
                 size_t iteration,
